Clean: Decoupled asset loaders from mesh header
